Created Keyboard Save
Created and implemented a keyboard save and load system. You press F5 to save and F9 to load the save. This is a basic save and load system, might be changed in the future. Also made several variables public within the weather manager script as the save script could not access them.